requirements.txt
asgiref==3.2.10
astroid==2.4.2
boto3==1.14.17
botocore==1.17.17
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.7
django-admin-views==0.8.0
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-sslserver==0.22
django-storages==1.9.1
django-tags-input==5.0.0
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.20
pylint==2.5.3
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
insta-scrape == 2.1.2